+++ crates/openehr-rm/src/model/data.rs
@@ -8,15 +8,6 @@ use super::{
 
 pub(super) static CLASSES: &[RmClass] = &[
     RmClass {
-        name: "ACCESS_CONTROL_SETTINGS",
-        is_abstract: true,
-        ancestors: &[],
-        descendants: &[],
-        attributes: &[],
-        is_structure_root: false,
-        generic_params: &[],
-    },
-    RmClass {
         name: "ACCESS_GROUP_REF",
         is_abstract: false,
         ancestors: &["OBJECT_REF"],